Allow to specify buffer timeout via PasteBufferTimeout. Use a timer.
See matterircd.toml.example for an example.

PasteBufferTimeout specifies the amount of time in milliseconds that
messages get kept in matterircd internal buffer before being sent to
mattermost or slack.
Messages that will be received in this time will be concatenated together
So this can be used to paste stuff like ansi-art or code.
Default 0 (is disabled)
Depending on how fast you type 2500 is a good number