Allow binding to a Unix socket (#276)
For a single-user bridge, listening on a Unix socket can be more
convenient than listening on a TCP port: you don't have to worry about
port security or port clashes (especially if running on somebody else's
multi-user system), and you can just rely on filesystem permissions on
the socket in question.  Of course you need an IRC client that supports
connecting to Unix sockets, but e.g. irssi does.

If the bind option is given a path (detected by checking whether it
contains the path separator character), treat it as the path to a Unix
socket and bind to that.