Add support for deleting/modifying own messages (mattermost)
Also needs the prefixcontext option.
With this feature you can modify or delete messages by doing

- to modify: s/XXX/new text
- to delete: s/XXX/

where XXX is a hexnumber between 000 and fff which you have
to calculate yourself based on the context of the other users
messages.

See example, we now our message will be 003 so we can change it
and on the mattermost side, it'll be the new text

```
23:25 <@wim> [001] hi
23:25 <@wim> [002] something
23:25 < wimirc> hllo how are you
23:25 < wimirc> s/003/hello how are you
23:25 <@wim> [004] fine
```

Also adds changes to the bridge interface as we now need to receive
msgID's when we send messages.