Fix buffer flushing yet again (#411)
Yet another fix to bug introduced by commit bc88bb1. Still an issue
with pasting paths causing the buffers to be flushed when it shouldn't
(e.g. /srv/mydir). Instead, split it into two regexs to avoid
confusion. One for replies to message/threads as well as reactions and
the other for message modifications.