Support saving last viewed at state for multiple users (mattermost) (#425)
Persistent last viewed at state was added in PR #313. It is done by
dumping out the state of the current user to a file on
disk. Unfortunately, this is only per user, the current user, even
though matterircd supports multiple users.

This fixes that by switching to using bbolt for storing and saving the
last viewed at state.