Fix incorrect msg/thread counters (#542)
When using mattericd's default threading, the msgMap is populated with
message/thred IDs and never removed. This means that when we cycle
past 4k, we can have entries with duplicate counters causing issues
when replying to threads.

This fixes it by introducing an index.

Also, rework the way Matterircd threading works.